+++ src/helics/application_api/Federate.cpp
@@ -1012,12 +1012,18 @@ std::string Federate::query(const std::string& queryStr)
     std::string res;
     if (queryStr == "name") {
         res = getName();
-    } else if (queryStr == "corename") {
+    }
+    else if (queryStr == "corename") {
         if (coreObject) {
             res = coreObject->getIdentifier();
-        } else {
+        }
+        else {
             res = "#unknown";
         }
+    }
+    else if (queryStr == "current_time")
+    {
+        res = std::to_string(currentTime);
     } else {
         res = localQuery(queryStr);
     }